# openapi/api.yaml
openapi: 3.0.0
info:
  title: PianoML API
  version: 1.0.0
paths:
  /account/login:
    post:
      tags:
        - account
      summary: Log in to an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  username:
                    type: string
                  userId:
                    type: string
                  expire:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
  /account/create:
    post:
      operationId: accountCreatePost
      tags:
        - account
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                name:
                  type: string
              required:
                - email
                - password
                - name
      responses:
        '201':
          description: Account created successfully
        '400':
          description: Bad request
        '403':
          description: User already exists
  /account/userinfo:
    get:
      operationId: accountUserinfoGet
      tags:
        - account
      summary: Get user information
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiInfo'
        '401':
          description: Unauthorized
    put:
      operationId: accountUserinfoPut
      tags:
        - account
      summary: Update user information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserApiInfo'
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiInfo'
        '401':
          description: Unauthorized
  /account/password_reset:
    post:
      operationId: accountPasswordResetPost
      tags:
        - account
      summary: Reset account password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        '200':
          description: Password reset email sent
        '400':
          description: Bad request
  /account/create/{token}/confirm:
    get:
      operationId: accountCreateTokenConfirmGet
      tags:
        - account
      summary: Confirm account creation with token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: The token to confirm account creation
      responses:
        '200':
          description: Account confirmed successfully
        '400':
          description: Invalid or expired token
  /account/logout:
    get:
      operationId: accountLogoutGet
      tags:
        - account
      summary: Log out from the account
      responses:
        '200':
          description: Successfully logged out
        '401':
          description: Unauthorized

  /score:
    post:
      operationId: scorePost
      tags:
        - score
      summary: Create a score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreApiInfo'
      responses:
        '200':
          description: ScoreApiInfo successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreApiInfo'
        '400':
          description: Bad request

  /score/search:
    get:
      operationId: scoreSearchGet
      tags:
        - score
      summary: Search a score
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
          description: A keyword
        - name: ownerId
          in: query
          schema:
            type: string
          description: owner
        - name: genreId
          in: query
          schema:
            type: string
          description: genre
        - name: artist
          in: query
          schema:
            type: string
          description: artist name
        - name: etude
          in: query
          schema:
            type: boolean
          description: filter by etude type
        - name: gradeStart
          in: query
          schema:
            type: integer
        - name: gradeEnd
          in: query
          schema:
            type: integer
          description: the level
        - name: offset
          in: query
          schema:
            type: integer
          description: search start at
        - name: limit
          in: query
          schema:
            type: integer
          description: limit results count
      responses:
        '200':
          description: ScoreApiInfo retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreApiInfo'
        '400':
          description: Bad request

  /score/{id}/info:
    get:
      operationId: scoreIdGet
      tags:
        - score
      summary: Retrieve a score by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the score to retrieve
      responses:
        '200':
          description: ScoreApiInfo retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreApiInfo'
        '404':
          description: ScoreApiInfo not found
    put:
      operationId: scoreIdPut
      tags:
        - score
      summary: Update a score by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the score to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreApiInfo'

      responses:
        '200':
          description: ScoreApiInfo retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreApiInfo'
        '403':
          description: Not authorized
        '404':
          description: ScoreApiInfo not found
    delete:
      operationId: scoreIdDelete
      tags:
        - score
      summary: Delete a score by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the score to delete
      responses:
        '204':
          description: Score deleted successfully
        '403':
          description: Not authorized - only owner or admin can delete
        '404':
          description: Score not found

  /score/{slug}:
    get:
      operationId: scoreGetBySlug
      tags:
        - score
      summary: Retrieve a score by its immutable slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The immutable slug of the score to retrieve
      responses:
        '200':
          description: Score retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreApiInfo'
        '404':
          description: Score not found

  /score/{mbid}/{type}/{version}/{revision}:
    post:
      operationId: scoreMbidTypeVersionRevisionPost
      tags:
        - score
      summary: Upload a binary file for a specific score
      parameters:
        - name: mbid
          in: path
          required: true
          schema:
            type: string
          description: The ID of the score
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [pdf, musicxml, midi, metadata]
          description: The type of file being uploaded (e.g., pdf or mxml)
        - name: version
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the score
        - name: revision
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the score
        - name: track1
          in: query
          required: false
          schema:
            type: integer
        - name: track2
          in: query
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: File uploaded successfully
        '400':
          description: Invalid input
        '404':
          description: ScoreApiInfo not found
  /score/{owner}/{mbid}/{type}/{version}/{revision}:
    get:
      operationId: scoreOwnerMbidTypeVersionRevisionGet
      tags:
        - score
      summary: Download a file for a specific score
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
          description: The User Id
        - name: mbid
          in: path
          required: true
          schema:
            type: string
          description: The ID of the score
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [pdf, musicxml, midi]
          description: The type of file to download (e.g., pdf or mxml)
        - name: version
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the score
        - name: revision
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the score

      responses:
        '200':
          description: File downloaded successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: ScoreApiInfo or file type not found

  /workload/{id}:
    get:
      operationId: workloadIdGet
      tags:
        - workload
      summary: Retrieve a workload by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the workload to retrieve
      responses:
        '200':
          description: Workload retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkloadApiInfo'
        '404':
          description: Workload not found

  /author:
    post:
      operationId: authorPost
      tags:
        - author
      summary: Create a new author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorApiInfo'

      responses:
        '201':
          description: AuthorApiInfo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorApiInfo'
        '400':
          description: Invalid input
  /author/{mbid}:
    put:
      operationId: authorMbidPut
      tags:
        - author
      summary: Update an author
      parameters:
        - name: mbid
          in: path
          required: true
          schema:
            type: string
          description: The MusicBrainz ID for the author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorApiInfo'
      responses:
        '200':
          description: AuthorApiInfo retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorApiInfo'
        '404':
          description: AuthorApiInfo not found
    get:
      operationId: authorMbidGet
      tags:
        - author
      summary: Retrieve informations about an author
      parameters:
        - name: mbid
          in: path
          required: true
          schema:
            type: string
          description: The MusicBrainz ID for the author
      responses:
        '200':
          description: AuthorApiInfo retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorApiInfo'
        '404':
          description: AuthorApiInfo not found
  /author/search/{query}:
    get:
      operationId: authorSearchQueryGet
      tags:
        - author
      summary: Search authors
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
          description: The ID of the author
      responses:
        '200':
          description: List of authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorApiInfo'
        '404':
          description: AuthorApiInfo not found
  /genre:
    get:
      operationId: genreGet
      tags:
        - genre
      summary: Retrieve all genres
      responses:
        '200':
          description: A list of genres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenreApiInfo'
    post:
      operationId: genrePost
      tags:
        - genre
      summary: Add a genre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreApiInfo'

      responses:
        '201':
          description: Genres created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreApiInfo'
  /genre/{id}:
    put:
      operationId: genreIdPut
      tags:
        - genre
      summary: Update a genre
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the genre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreApiInfo'
      responses:
        '200':
          description: Genre updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreApiInfo'
        '404':
          description: GenreApiInfo not found
    get:
      operationId: genreIdGet
      tags:
        - genre
      summary: Retrieve informations about a genre
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the genre
      responses:
        '200':
          description: Genre retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreApiInfo'
        '404':
          description: GenreApiInfo not found
  /genre/search/{query}:
    get:
      operationId: genreSearchQueryGet
      tags:
        - genre
      summary: Search genres
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
          description: The query to search genres
      responses:
        '200':
          description: List of genres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenreApiInfo'
        '404':
          description: Genre not found
  /artist/search/{query}:
    get:
      operationId: artistSearchQueryGet
      tags:
        - musicbrainz
      summary: Search all artists
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: AuthorApiInfo retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MbAuthorApiInfo'
        '404':
          description: GenreApiInfo not found

  /work/search/{query}:
    get:
      operationId: workSearchQueryGet
      tags:
        - musicbrainz
      summary: Search all works
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: AuthorApiInfo retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllWorksApiInfo'
        '404':
          description: Work not found


components:
  schemas:
    AuthorApiInfo:
      type: object
      properties:
        id:
          type: string
        mbid:
          type: string
        name:
          type: string
        disambiguation:
          type: string
        country:
          type: string
        type:
          type: string
        gender:
          type: string
        lifeSpanBegin:
          type: string
          format: date
        lifeSpanEnd:
          type: string
          format: date
        lifeSpanEnded:
          type: boolean
    ScoreApiInfo:
      type: object
      properties:
        id:
          type: string
        mbid:
          type: string
        title:
          type: string
        author:
          type: string
        author_id:
          type: string
        sortName:
          type: string
        version:
          type: integer
          format: int32
        tracks_count:
          type: integer
          format: int32
        hand_separated:
          type: boolean
        has_lyrics:
          type: boolean
        measures:
          type: integer
          format: int32
        duration:
          type: number
          format: float
        genre:
          type: string
        genre_id:
          type: string
        grade:
          type: integer
          format: int32
        uploaded_at:
          type: string
          format: date-time
        owner_id:
          type: string
        owner:
          type: string
        updated_at:
          type: string
          format: date-time
        has_files:
          type: boolean
        image:
          type: string
          format: uri
        study_tracks:
          type: array
          items:
            type: integer
        mutableSlug:
          type: string
        immutableSlug:
          type: string

    WorkloadApiInfo:
      type: object
      properties:
        id:
          type: integer
          format: int32
        kind:
          type: string
        createdAt:
          type: string
          format: date-time
        scoreId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        errorMessage:
          type: string
        duration:
          type: integer
          format: int32
          description: Duration in milliseconds
        workloadSize:
          type: integer
          format: int32

    UserApiInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        url:
          type: string
        bio:
          type: string
          format: uri
        image:
          type: string
          format: uri
      required:
        - name
    GenreApiInfo:
      title: GenreApiInfo
      type: object
      properties:
        id:
          type: string
        mbid:
          type: string
        name:
          type: string
    MusicBrainzService:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        disambiguation:
          type: string
        sort-name:
          type: string
        type:
          type: string
        life-span:
          type: object
          properties:
            begin:
              type: string
              format: date
            end:
              type: string
              format: date
            ended:
              type: boolean
    AllWorksApiInfo:
      type: object
      properties:
        works:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              relations:
                type: array
                items:
                  type: object
                  properties:
                    artist:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        sort-name:
                          type: string
                        disambiguation:
                          type: string
    MbAuthorApiInfo:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
        sort-name:
          type: string
        disambiguation:
          type: string
        gender:
          type: string
        country:
          type:  string
        life-span:
          type: object
          properties:
            begin:
              type: string
              format: date
            end:
              type: string
              format: date
            ended:
              type: boolean
    ErrorApiInfo:
      type: object
      properties:
        error:
          type: object
          properties:
            name:
              type: string
            message:
              type: string
            stacktrace:
              type: array
              items:
                type: string